Update to fix several games
- Added SetFvF to initialize vertex declaration to fix an issue with Max
Payne.
- Removed SetSoftwareVertexProcessing() to fix an issue with Haegemonia:
Legions of Iron.
- Modified the code that translates the vertex declaration so that if
the Usage is blendindices it will set the Method to Default.  This fixes
an error in CreateVertexDeclaration found in True Crime New York City.
- Updated the pixel shader translation to ignore co-issued statements to
resolve an issue found in True Crime New York City.
- Updated the pixel shader translation to ensure that double swizzles
are not added when both the modified and modifier registers both have
swizzles.  This fixes an issue found in True Crime New York City.